# specs/002-fullstack-web-todo/contracts/tasks.openapi.yaml
openapi: 3.0.0
info:
  title: Todo API - Phase II Tasks
  version: 2.0.0
  description: API for managing Todo tasks, featuring multi-user support and JWT authentication.

servers:
  - url: http://localhost:8000/api
    description: Local Development Server

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskBase:
      type: object
      properties:
        title:
          type: string
          description: The title of the task
          minLength: 1
        description:
          type: string
          description: Optional detailed description of the task
          nullable: true
        complete:
          type: boolean
          description: Whether the task is completed or not
          default: false
      required:
        - title

    TaskCreate:
      allOf:
        - $ref: '#/components/schemas/TaskBase'

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          description: The updated title of the task
          minLength: 1
        description:
          type: string
          description: The updated detailed description of the task
          nullable: true
        complete:
          type: boolean
          description: Whether the task is completed or not
      minProperties: 1

    TaskPatchComplete:
      type: object
      properties:
        complete:
          type: boolean
          description: The new completion status of the task
      required:
        - complete

    TaskRead:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: Unique identifier for the task
            user_id:
              type: string
              format: uuid
              description: ID of the user who owns the task
            created_at:
              type: string
              format: date-time
              description: Timestamp when the task was created
            updated_at:
              type: string
              format: date-time
              description: Timestamp when the task was last updated
          required:
            - id
            - user_id
            - created_at
            - updated_at

    HTTPValidationError:
      type: object
      title: HTTPValidationError
      properties:
        detail:
          type: array
          title: Detail
          items:
            $ref: '#/components/schemas/ValidationError'

    ValidationError:
      type: object
      title: ValidationError
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          title: Location
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type

paths:
  /tasks:
    get:
      summary: Get all tasks for the authenticated user
      operationId: get_tasks
      tags:
        - Tasks
      responses:
        '200':
          description: A list of tasks for the current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

    post:
      summary: Create a new task for the authenticated user
      operationId: create_task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: The newly created task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

  /tasks/{task_id}:
    get:
      summary: Get a single task by ID for the authenticated user
      operationId: get_task_by_id
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the task to retrieve
      responses:
        '200':
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '404':
          description: Task not found for the current user.

    put:
      summary: Update an existing task by ID for the authenticated user
      operationId: update_task
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: The updated task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '404':
          description: Task not found for the current user.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

    delete:
      summary: Delete a task by ID for the authenticated user
      operationId: delete_task
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the task to delete
      responses:
        '204':
          description: Task successfully deleted (No Content).
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '404':
          description: Task not found for the current user.

  /tasks/{task_id}/complete:
    patch:
      summary: Mark a task as complete or incomplete for the authenticated user
      operationId: patch_task_complete
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the task to update completion status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskPatchComplete'
      responses:
        '200':
          description: The updated task with new completion status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '404':
          description: Task not found for the current user.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'